refactor: 점수 계산 로직을 BoardGame에게 위임
