+++ src/main/java/woowa/camp/chess/BoardGame.java
@@ -1,16 +1,29 @@
 package woowa.camp.chess;
 
+import static woowa.camp.chess.BoardConstants.MAX_COL;
 import static woowa.camp.pieces.Piece.Type.NO_PIECE;
 
+import java.util.List;
 import woowa.camp.pieces.Piece;
+import woowa.camp.pieces.Piece.Color;
+import woowa.camp.pieces.Piece.Type;
 
 public class BoardGame {
 
     private final Board board;
 
-    public BoardGame(final Board board) {
-        board.initialize();
+    private BoardGame(final Board board) {
         this.board = board;
+    }
+
+    public static BoardGame createWithInitialize(final Board board) {
+        board.initialize();
+        return new BoardGame(board);
+    }
+
+    public static BoardGame createWithEmptyInitialize(final Board board) {
+        board.initializeEmpty();
+        return new BoardGame(board);
     }
 
     public void move(final String source, final String destination) {
@@ -28,6 +41,15 @@ public class BoardGame {
     private void doMove(final Piece sourcePiece, final Position destinationPosition, final Position sourcePosition) {
         board.replace(sourcePiece, destinationPosition);
         board.remove(sourcePosition);
+    }
+
+    public double calculateScore(final Color color) {
+        double score = 0.0;
+        for (int col = 0; col < MAX_COL.getCount(); col++) {
+            final List<Piece> file = board.extractPiecesByFile(col, color);
+            score += Type.calculateScore(file);
+        }
+        return score;
     }
 
     private void validateMove(final Piece sourcePiece, final Position sourcePosition, final Piece destinationPiece) {
